python-app/requirements.txt
grpcio==1.59.3
grpcio-reflection==1.59.3
